Fix Numbers

+++ GAZT/GAZT/Models/EDeclerationsModel/SubmitModels/EDeclerationSubmitModel.cs
@@ -22,7 +22,7 @@ namespace EGAZT.Models.EDeclerationsModel.SubmitModels
         public string typeName { get; set; }
         public string itemCode { get; set; }
         public int count { get; set; }
-        public int? value { get; set; }
+        public double? value { get; set; }
         public Guid ID { get; set; } = Guid.NewGuid();
     }
 
@@ -33,7 +33,7 @@ namespace EGAZT.Models.EDeclerationsModel.SubmitModels
         public string otherpurpose { get; set; }
         public int count { get; set; }
         public int unit { get; set; }
-        public int? value { get; set; }
+        public double? value { get; set; }
         public int currency { get; set; }
         public string currencyName { get; set; }
         public bool permit { get; set; }
